+++ packages/compiler-sfc/src/genCssVars.ts
@@ -38,7 +38,7 @@ export function genCssVarsCode(
           })
           .join('')
 
-  return `__useCSSVars__(_ctx => (${transformedString})${
+  return `__useCssVars__(_ctx => (${transformedString})${
     scoped ? `, true` : ``
   })`
 }
@@ -65,7 +65,7 @@ export function injectCssVarsCalls(
 
   return (
     script +
-    `\nimport { useCSSVars as __useCSSVars__ } from 'vue'\n` +
+    `\nimport { useCssVars as __useCssVars__ } from 'vue'\n` +
     `const __injectCSSVars__ = () => {\n${calls}}\n` +
     `const __setup__ = __default__.setup\n` +
     `__default__.setup = __setup__\n` +